requirements.txt
langchain==1.2.15
langchain-openai==1.2.0
langchain-community==0.4.1
langchain-chroma==1.1.0
langgraph==1.1.9
openai==2.32.0
chromadb==1.5.8
redis==7.4.0
tiktoken==0.12.0
pandas==3.0.2
tabulate==0.10.0
python-dotenv==1.2.2
pytest==8.3.5
